Poprawione rozwiązania zadań z zajęć 2

+++ w2/ZuberRS03/Zadanie7.java
@@ -11,7 +11,7 @@ public class Zadanie7 {
             suma = suma + n;
             x++;
         }while(n <= 100);
-        float srednia = suma/x;
+        float srednia = (float)suma/x;
         System.out.println("Średnia podanych liczb: " + srednia);
     }
 }
